refactor: Moved conversation logic to router
